requirements.txt
# Core dependencies
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-generativeai==0.7.2
yt-dlp==2024.8.6
pytube==15.0.0
moviepy==1.0.3

# NLP and text processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.4
textblob==0.18.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Logging and error handling
google-api-core==2.11.1

# Audio processing
ffmpeg-python==0.2.0

# Date and time handling
pytz==2024.1

# Additional dependencies
numpy==1.26.4
pandas==2.2.2

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
pytest-xdist==3.5.0
pytest-timeout==2.2.0
coverage==7.4.1
tox==4.12.1
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2